requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.2.0 